Canvas templates via templatesSet (#2014)
* wip: migration to create canvas templates based on canvases at any level that are templates

* updated to get canvas checkout directly from teh lifecycle; completed migration for both removing cached status + using canvasTemplate entity

* removed isTemplate field

* removed migration code related to status on canvas checkout

* added a real canvas into create canvas template sample

* renamed migration to be after other recent migrations

* updated to reflect displayName on canvas

* remove canvas from ecosystem model as not used

* remove old ecosystem model canvases

* remove index on ecosystemModel related to the canvas

* updated dropping of index

* remove FK

* added index into down migration

* only rename existing templates

* migration running

* allow specifying a canvas to pick up the value from when creating a canvas Template
