+++ graphql-samples/mutations/create/create-canvas-template
@@ -15,6 +15,6 @@ Variables:
         "description": "wow this is great",
         "tags": ["Tag 1", "Tag 2"]
       },
-      "value": "wow3"
+      "value": "{\n  \"type\": \"excalidraw\",\n  \"version\": 2,\n  \"source\": \"https://hub.alkem.io\",\n  \"elements\": [\n    {\n      \"id\": \"bXNwShWhTQXUyrHzyRz8X\",\n      \"type\": \"ellipse\",\n      \"x\": 848.3664283752441,\n      \"y\": 201.3920440673828,\n      \"width\": 289.09088134765625,\n      \"height\": 278.18182373046875,\n      \"angle\": 0,\n      \"strokeColor\": \"#000000\",\n      \"backgroundColor\": \"transparent\",\n      \"fillStyle\": \"hachure\",\n      \"strokeWidth\": 1,\n      \"strokeStyle\": \"solid\",\n      \"roughness\": 1,\n      \"opacity\": 100,\n      \"groupIds\": [],\n      \"strokeSharpness\": \"sharp\",\n      \"seed\": 1784787662,\n      \"version\": 23,\n      \"versionNonce\": 573886350,\n      \"isDeleted\": false,\n      \"boundElements\": null,\n      \"updated\": 1656182212930,\n      \"link\": null\n    },\n    {\n      \"id\": \"ivwQMjgRMAKImMISfYaJt\",\n      \"type\": \"rectangle\",\n      \"x\": 923.8210182189941,\n      \"y\": 262.30113220214844,\n      \"width\": 154.54547119140625,\n      \"height\": 149.99996948242188,\n      \"angle\": 0,\n      \"strokeColor\": \"#000000\",\n      \"backgroundColor\": \"transparent\",\n      \"fillStyle\": \"hachure\",\n      \"strokeWidth\": 1,\n      \"strokeStyle\": \"solid\",\n      \"roughness\": 1,\n      \"opacity\": 100,\n      \"groupIds\": [],\n      \"strokeSharpness\": \"sharp\",\n      \"seed\": 1676425298,\n      \"version\": 46,\n      \"versionNonce\": 871046990,\n      \"isDeleted\": false,\n      \"boundElements\": null,\n      \"updated\": 1656182216500,\n      \"link\": null\n    },\n    {\n      \"id\": \"OpXnwCwhFkAFLwZkmMi3s\",\n      \"type\": \"diamond\",\n      \"x\": 971.0937232971191,\n      \"y\": 281.3920440673828,\n      \"width\": 53.63641357421875,\n      \"height\": 113.6363525390625,\n      \"angle\": 0,\n      \"strokeColor\": \"#000000\",\n      \"backgroundColor\": \"transparent\",\n      \"fillStyle\": \"hachure\",\n      \"strokeWidth\": 1,\n      \"strokeStyle\": \"solid\",\n      \"roughness\": 1,\n      \"opacity\": 100,\n      \"groupIds\": [],\n      \"strokeSharpness\": \"sharp\",\n      \"seed\": 726910098,\n      \"version\": 98,\n      \"versionNonce\": 377599570,\n      \"isDeleted\": false,\n      \"boundElements\": null,\n      \"updated\": 1656182224691,\n      \"link\": null\n    }\n  ],\n  \"appState\": {\n    \"gridSize\": null,\n    \"viewBackgroundColor\": \"#ffffff\"\n  },\n  \"files\": {}\n}"
   }
 }
