+++ graphql-samples/mutations/create/create-canvas-template-from-canvas
@@ -1,0 +1,20 @@
+mutation createCanvasTemplate($canvasTemplateInput: CreateCanvasTemplateOnTemplatesSetInput!) {
+  createCanvasTemplate(canvasTemplateInput: $canvasTemplateInput) {
+    info {
+      title
+    }
+  }
+}
+
+Variables:
+{
+  "canvasTemplateInput": {
+      "templatesSetID": "uuid",
+      "info": {
+        "title": "wow",
+        "description": "wow this is great",
+        "tags": ["Tag 1", "Tag 2"]
+      },
+      "canvasID": "uuid"
+  }
+}
