+++ src/domain/community/organization/organization.resolver.fields.ts
@@ -3,8 +3,7 @@ import { Args, Context, Resolver } from '@nestjs/graphql';
 import { Parent, ResolveField } from '@nestjs/graphql';
 import { Organization } from './organization.entity';
 import { OrganizationService } from './organization.service';
-import { EntityNotInitializedException } from '@common/exceptions';
-import { AuthorizationPrivilege, LogContext } from '@common/enums';
+import { AuthorizationPrivilege } from '@common/enums';
 import { GraphqlGuard } from '@core/authorization';
 import { IOrganization } from '@domain/community/organization';
 import { IUserGroup } from '@domain/community/user-group';
