+++ src/domain/context/ecosystem-model/ecosystem-model.interface.ts
@@ -1,4 +1,3 @@
-import { ICanvas } from '@domain/common/canvas';
 import { IAuthorizable } from '@domain/common/entity/authorizable-entity';
 import { IActorGroup } from '@domain/context/actor-group/actor-group.interface';
 import { Field, ObjectType } from '@nestjs/graphql';
@@ -16,8 +15,6 @@ export abstract class IEcosystemModel extends IAuthorizable {
     description: 'A list of ActorGroups',
   })
   actorGroups?: IActorGroup[];
-
-  canvas?: ICanvas;
 
   restrictedActorGroupNames!: string[];
 }
