+++ src/domain/context/ecosystem-model/ecosystem-model.service.ts
@@ -21,7 +21,6 @@ import { ActorGroupService } from '@domain/context/actor-group/actor-group.servi
 import { AuthorizationPolicy } from '@domain/common/authorization-policy';
 import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
 import { CanvasService } from '@domain/common/canvas/canvas.service';
-import { Canvas, ICanvas } from '@domain/common/canvas';
 
 @Injectable()
 export class EcosystemModelService {
@@ -41,7 +40,6 @@ export class EcosystemModelService {
     ecosystemModel.authorization = new AuthorizationPolicy();
     await this.createRestrictedActorGroups(ecosystemModel);
     ecosystemModel.actorGroups = [];
-    // ecosystemModel.canvas = new Canvas();
     return await this.ecosystemModelRepository.save(ecosystemModel);
   }
 
@@ -64,12 +62,6 @@ export class EcosystemModelService {
     ecosystemModelInput: UpdateEcosystemModelInput
   ): Promise<IEcosystemModel> {
     ecosystemModel.description = ecosystemModelInput.description;
-    if (ecosystemModelInput.canvas) {
-      ecosystemModel.canvas = this.canvasService.updateCanvasEntity(
-        ecosystemModel.canvas,
-        ecosystemModelInput.canvas
-      );
-    }
     return await this.ecosystemModelRepository.save(ecosystemModel);
   }
 
@@ -91,9 +83,6 @@ export class EcosystemModelService {
       await this.authorizationPolicyService.delete(
         ecosystemModel.authorization
       );
-
-    if (ecosystemModel.canvas)
-      await this.canvasService.deleteCanvas(ecosystemModel.canvas.id);
 
     return await this.ecosystemModelRepository.remove(
       ecosystemModel as EcosystemModel
@@ -164,14 +153,5 @@ export class EcosystemModelService {
         LogContext.CONTEXT
       );
     return actorGroups;
-  }
-
-  async getCanvas(ecosystemModel: IEcosystemModel): Promise<ICanvas> {
-    if (!ecosystemModel.canvas) {
-      // create and add the canvas
-      ecosystemModel.canvas = new Canvas();
-      await this.ecosystemModelRepository.save(ecosystemModel);
-    }
-    return ecosystemModel.canvas;
   }
 }
