+++ src/domain/template/canvas-template/dto/canvas.template.dto.create.ts
@@ -1,4 +1,8 @@
-import { CANVAS_VALUE_LENGTH } from '@common/constants/entity.field.length.constants';
+import {
+  CANVAS_VALUE_LENGTH,
+  UUID_LENGTH,
+} from '@common/constants/entity.field.length.constants';
+import { UUID } from '@domain/common/scalars/scalar.uuid';
 import { CreateTemplateBaseInput } from '@domain/template/template-base/dto';
 import { Field, InputType } from '@nestjs/graphql';
 import { IsOptional, MaxLength } from 'class-validator';
@@ -11,4 +15,13 @@ export class CreateCanvasTemplateInput extends CreateTemplateBaseInput {
   @IsOptional()
   @MaxLength(CANVAS_VALUE_LENGTH)
   value?: string;
+
+  @Field(() => UUID, {
+    nullable: true,
+    description:
+      'Use the specified Canvas as the initial value for this CanvasTempplate',
+  })
+  @IsOptional()
+  @MaxLength(UUID_LENGTH)
+  canvasID?: string;
 }
